Move `mapNorisToTaxDetailData` logic from `tax-detail.helper.ts` to `noris.helper.ts`.
