+++ nest-tax-backend/src/admin/admin.service.ts
@@ -35,10 +35,10 @@ import {
   convertCurrencyToInt,
   mapNorisToTaxAdministratorData,
   mapNorisToTaxData,
+  mapNorisToTaxDetailData,
   mapNorisToTaxInstallmentsData,
   mapNorisToTaxPayerData,
 } from '../noris/utils/noris.helper'
-import { mapNorisToTaxDetailData } from '../noris/utils/tax-detail.helper'
 import { createTestingTaxMock } from './utils/testing-tax-mock'
 
 @Injectable()
